Adicionando as classes Service e RestController para a entidade Usuario

+++ java/seboDoJoao/src/main/java/br/com/interdisciplinar/sebodojoao/repository/CidadeRepository.java
@@ -3,5 +3,5 @@ package br.com.interdisciplinar.sebodojoao.repository;
 import br.com.interdisciplinar.sebodojoao.model.Cidade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CidadeRepository extends JpaRepository<Cidade, Integer> {
+public interface CidadeRepository extends JpaRepository<Cidade, Long> {
 }
